requirements.txt
Django==3.1
djangorestframework==3.11.1
Pint==0.14